requirements.txt
asgiref==3.2.3
Django==3.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
